requirements.txt
# Core AI APIs
anthropic==0.7.8
google-generativeai==0.3.2
elevenlabs

# Google Cloud Services
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread==5.12.0

# HTTP clients
httpx
requests==2.31.0

# Audio processing
pydub==0.25.1

# Video processing
ffmpeg-python==0.2.0

# Image processing
Pillow==10.1.0

# Text processing
python-slugify==8.0.1
rapidfuzz==3.5.2

# Data handling
pandas==2.1.3

# Environment management
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-xdist==3.5.0
ruff==0.1.2

# Utilities
psutil==5.9.6

# Additional production dependencies
gunicorn==21.2.0
flask==3.0.0



